Refactoring server structure
