+++ package.json
@@ -18,7 +18,6 @@
   "dependencies": {
     "classnames": "2.2.6",
     "core-js": "2.5.7",
-    "cors": "2.8.4",
     "deepmerge": "2.2.1",
     "express": "4.16.4",
     "express-winston": "3.0.1",
@@ -54,7 +53,6 @@
     "@types/autoprefixer": "^9.1.0",
     "@types/case-sensitive-paths-webpack-plugin": "^2.1.2",
     "@types/core-js": "^2.5.0",
-    "@types/cors": "^2.8.4",
     "@types/dotenv": "^4.0.3",
     "@types/enzyme": "^3.1.0",
     "@types/enzyme-adapter-react-16": "^1.0.3",
