+++ src/server/middleware/index.ts
@@ -1,0 +1,4 @@
+/**
+ * Register Middleware here
+ */
+export { default as renderer } from "./renderer";
