+++ src/server/middleware/renderer.tsx
@@ -1,30 +1,31 @@
-import {Request, Response} from "express";
+import {NextFunction, Request, Response} from "express";
 import * as React from "react";
 import {renderToNodeStream} from "react-dom/server";
 import {matchPath, RouteProps, StaticRouter} from "react-router";
-import {Stats} from "webpack";
-import App from "../shared/App";
-import routes from "../shared/routes";
-import {Html} from "./templates";
+import App from "../../shared/App";
+import routes from "../../shared/routes";
+import {Html} from "../views";
 
-interface RendererProps {
-    clientStats: Stats;
-    serverStats: Stats;
-}
-
-export default (options: RendererProps) => (req: Request, res: Response): void => {
+/**
+ * Example middleware
+ * @param req
+ * @param res
+ * @param next
+ */
+export default (req: Request, res: Response, next: NextFunction): void => {
     const context = {};
     const activeRoute = routes.find((route: RouteProps) => !!matchPath(req.url, route));
     if (activeRoute) {
         // `<!doctype html>` add tp stream???
         renderToNodeStream(
-            <Html>
+            <Html files={req.files}>
                 <StaticRouter location={req.url} context={context}>
                     <App />
                 </StaticRouter>
             </Html>,
         ).pipe(res);
     } else {
-        res.status(404).send("O no we should be serving a 404!");
+        res.status(404).send("O no we should be serving a 404!"); // TODO handle this in react and then pull back
     }
+    // TODO what about next()???
 };
